broke down form into small components
